Revert "Revert "Refactor the handler call graph to attach constructors and input parameters to nodes instead of types.""
This reverts commit 2e5cb0965e935ff1c2cc3bab41bb6848685f9f64.
