+++ libs/pavex_cli/tests/ui_tests/multiple_versions_for_the_same_crate_are_supported/expectations/diagnostics.dot
@@ -1,7 +1,7 @@
 digraph "/home" {
     0 [ label = "app::stream_file(http_1::header::HeaderName, http_0::header::HeaderName) -> http_0::Response<hyper::Body>"]
-    1 [ label = "http_0::header::HeaderName"]
-    2 [ label = "http_1::header::HeaderName"]
+    1 [ label = "app::header2() -> http_0::header::HeaderName"]
+    2 [ label = "app::header1() -> http_1::header::HeaderName"]
     1 -> 0 [ ]
     2 -> 0 [ ]
 }
