+++ libs/pavex_cli/tests/ui_tests/references_to_constructible_types_are_allowed/expectations/diagnostics.dot
@@ -1,9 +1,9 @@
 digraph "/home" {
     0 [ label = "app::stream_file(&app::Singleton, &app::RequestScoped, &app::Transient) -> http::Response<hyper::Body>"]
     1 [ label = "&app::Transient"]
-    2 [ label = "app::Transient"]
+    2 [ label = "app::transient() -> app::Transient"]
     3 [ label = "&app::RequestScoped"]
-    4 [ label = "app::RequestScoped"]
+    4 [ label = "app::request_scoped() -> app::RequestScoped"]
     5 [ label = "&app::Singleton"]
     1 -> 0 [ ]
     2 -> 1 [ ]
