+++ libs/pavex_cli/tests/ui_tests/the_same_constructor_can_be_registered_multiple_times_using_the_same_lifecycle/expectations/diagnostics.dot
@@ -1,6 +1,6 @@
 digraph "/home" {
     0 [ label = "app::Streamer::stream_file(app::Logger) -> http::Response<hyper::Body>"]
-    1 [ label = "app::Logger"]
+    1 [ label = "app::new_logger() -> app::Logger"]
     1 -> 0 [ ]
 }
 digraph app_state {
